requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
h11==0.16.0
idna==3.11
openai
pinecone-client
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
PyGithub
